requirements.txt
python-telegram-bot==20.4
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
# pyppeteer==1.0.2  # Может вызывать проблемы в Railway контейнерах
pyppeteer==1.0.2
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
telethon==1.34.0
psutil==7.0.0
